+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 
-##2019/08/31
+## 2019/12/04
+
+- New option *singleGroupOpen* to allow only one group to be open and collapse all others.
+
+## 2019/08/31
 
 - Implementation of case-insensitive filter search as default with ability to override. Thanks to Mrkbingham.